refactor: FarePolicy의  enum별 요금 계산 방법을 인터페이스로 변경
