+++ src/main/java/wooteco/subway/domain/fare/farepolicy/BaseCalculateStrategy.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain.fare.strategy;
+package wooteco.subway.domain.fare.farepolicy;
 
 public class BaseCalculateStrategy implements FareCalculateStrategy {
